refactor: 원두 관련 기능을 beans 앱으로 이전
- profiles 앱에서 beans 앱으로 원두 관련 기능 이전
  - UserBeanListSchema를 beans/schemas.py로 이동
  - UserBeanListAPIView를 beans/views.py로 이동
  - get_user_saved_beans 로직을 BeanService 클래스로 이전

- URL 패턴 정리
  - /profiles/<id>/beans/ → /beans/<id>/beans/로 변경

- 필터 관련 코드 정리
  - records/filters.py 파일 제거
  - 필터 클래스들을 common/filters.py로 이동
