+++ repo/beans/schemas.py
@@ -1,7 +1,12 @@
 from drf_spectacular.types import OpenApiTypes
-from drf_spectacular.utils import OpenApiParameter, extend_schema
+from drf_spectacular.utils import (
+    OpenApiParameter,
+    OpenApiResponse,
+    extend_schema,
+    extend_schema_view,
+)
 
-from repo.beans.serializers import BeanSerializer
+from repo.beans.serializers import BeanSerializer, UserBeanSerializer
 from repo.common.serializers import PageNumberSerializer
 
 BeansTag = "beans"
@@ -25,3 +30,52 @@ class BeanSchema:
            """,
         tags=[BeansTag],
     )
+
+
+class UserBeanListSchema:
+    user_bean_list_get_schema = extend_schema(
+        responses={
+            200: UserBeanSerializer(many=True),
+            404: OpenApiResponse(description="Not Found"),
+        },
+        parameters=[
+            OpenApiParameter(name="bean_type", type=str, enum=["single", "blend"]),
+            OpenApiParameter(
+                name="origin_country",
+                type=str,
+                enum=[
+                    "케냐",
+                    "과테말라",
+                    "에티오피아",
+                    "브라질",
+                    "콜비아",
+                    "인도네시아",
+                    "온두라스",
+                    "탄자니아",
+                    "르완다",
+                ],
+                required=False,
+            ),
+            OpenApiParameter(name="is_decaf", type=bool, enum=[True, False], required=False),
+            OpenApiParameter(name="avg_star_min", type=float, enum=[x / 2 for x in range(11)], required=False),
+            OpenApiParameter(name="avg_star_max", type=float, enum=[x / 2 for x in range(11)], required=False),
+            OpenApiParameter(name="roast_point_min", type=int, enum=range(0, 6), required=False),
+            OpenApiParameter(name="roast_point_max", type=int, enum=range(0, 6), required=False),
+            OpenApiParameter(
+                name="ordering",
+                type=str,
+                enum=["-note__created_at", "-avg_star", "-tasted_records_cnt"],
+                required=False,
+            ),
+        ],
+        summary="유저 찜한 원두 리스트 조회",
+        description="""
+            특정 사용자가 저장한 원두 리스트를 조회합니다.
+            필터링: 원두 종류, 원산지, 디카페인 여부, 평균 별점, 로스팅
+            정렬: 노트 생성일, 평균 별점, 시음기록 수
+            담당자 : hwstar1204
+        """,
+        tags=[BeansTag],
+    )
+
+    user_bean_list_schema_view = extend_schema_view(get=user_bean_list_get_schema)
