+++ repo/beans/views.py
@@ -1,9 +1,12 @@
+from django_filters import rest_framework as filters
+from rest_framework import generics
 from rest_framework.pagination import PageNumberPagination
 from rest_framework.views import APIView
 
 from repo.beans.schemas import *
-from repo.beans.serializers import BeanSerializer
+from repo.beans.serializers import BeanSerializer, UserBeanSerializer
 from repo.beans.services import BeanService
+from repo.common.filters import BeanFilter
 
 
 @BeanSchema.bean_name_search_schema
@@ -22,3 +25,17 @@ class BeanNameSearchView(APIView):
 
         serializer = BeanSerializer(paginated_beans, many=True)
         return paginator.get_paginated_response(serializer.data)
+
+
+@UserBeanListSchema.user_bean_list_schema_view
+class UserBeanListAPIView(generics.ListAPIView):
+    serializer_class = UserBeanSerializer
+    filter_backends = [filters.DjangoFilterBackend]
+    filterset_class = BeanFilter
+    ordering_fields = ["-note__created_at", "-avg_star", "-tasted_records_cnt"]
+
+    def get_queryset(self):
+        user_id = self.kwargs.get("id")
+        ordering = self.request.query_params.get("ordering", "-note__created_at")
+        queryset = BeanService().get_user_saved_beans(user_id)
+        return queryset.order_by(ordering)
